+++ docutils/docutils/parsers/rst/directives/misc.py
@@ -233,6 +233,7 @@ class Include(Directive):
         settings._source = self.options['source']
         document = utils.new_document(settings._source, settings)
         document.include_log = self.state.document.include_log
+        document.names = self.state.document.names
         document.ids = self.state.document.ids
         document.nameids = self.state.document.nameids
         document.nametypes = self.state.document.nametypes